+++ cli/completion/bash/lnmp-docker.sh
@@ -19,6 +19,7 @@ complete -W "backup \
         push \
         restore \
         ssl \
+        ssl-self \
         swarm-build \
         swarm-push \
         swarm-deploy \
@@ -29,6 +30,7 @@ complete -W "backup \
         mongo-cli \
         mysql-cli \
         nginx-cli \
+        nginx-conf \
         php-cli \
         phpmyadmin-cli \
         postgres-cli \